Some refactor in prep for cli work.
